refactor: 写真確認画面をページからダイアログに変更
- PhotoConfirmPage を PhotoConfirmDialog に変換し、CameraPage 内で表示するように変更
- CameraPageArgs に onPhotoAccepted コールバックを追加し、採点処理を外部から注入できるように変更
- 撮影後の確認・採点ローディング表示を CameraPage の _handleCapture メソッドに集約
- /photo-confirm ルートを削除し、ルーティングをシンプル化
- mission_page.dart のカメラフロー制御を簡略化

Refs: #132
